+++ CHANGELOG.md
@@ -8,6 +8,7 @@ send sensitive data in logger (instead of private **_sendSensitive** parameter)
 * Logger events in debug mode won't be sent to remote logger anymore
 * Error in response will be logged in remote logger with response body despite
 **logSensitive** parameter
+* Network logger functions cleared
 
 
 ## 0.0.3